Switch creating dimensions to use templates
